examples/requirements.txt
fpdf
pdf2image
names
pyarrow
pandas
transformers
litdata
lightning
joblib
torchvision
scikit-learn